+++ example/a2dp_source_demo.c
@@ -200,18 +200,6 @@ static modcontext mod_context;
 static tracker_buffer_state trkbuf;
 
 /* AVRCP Target context START */
-static const uint8_t subunit_info[] = {
-    0,0,0,0,
-    1,1,1,1,
-    2,2,2,2,
-    3,3,3,3,
-    4,4,4,4,
-    5,5,5,5,
-    6,6,6,6,
-    7,7,7,7
-};
-
-static uint32_t company_id = 0x112233;
 static uint8_t  companies_num = 1;
 static uint8_t  companies[] = {
     0x00, 0x19, 0x58 //BT SIG registered CompanyID
@@ -765,9 +753,7 @@ static void avrcp_packet_handler(uint8_t packet_type, uint16_t channel, uint8_t 
             avrcp_subevent_connection_established_get_bd_addr(packet, event_addr);
 
             avrcp_target_set_now_playing_info(media_tracker.avrcp_cid, NULL, sizeof(tracks)/sizeof(avrcp_track_t));
-            avrcp_target_set_unit_info(media_tracker.avrcp_cid, AVRCP_SUBUNIT_TYPE_AUDIO, company_id);
-            avrcp_target_set_subunit_info(media_tracker.avrcp_cid, AVRCP_SUBUNIT_TYPE_AUDIO, (uint8_t *)subunit_info, sizeof(subunit_info));
-            
+
             avrcp_controller_get_supported_events(media_tracker.avrcp_cid);
             
             printf("AVRCP: Channel successfully opened:  media_tracker.avrcp_cid 0x%02x\n", media_tracker.avrcp_cid);